+++ src/filmgrain_tmpl_8.rs
@@ -683,7 +683,7 @@ unsafe extern "C" fn fgy_32x32xn_c(
     }
     let mut offsets: [[libc::c_int; 2]; 2] = [[0; 2]; 2];
     let mut bx: libc::c_uint = 0 as libc::c_int as libc::c_uint;
-    while (bx as libc::c_ulong) < pw {
+    while (bx as size_t) < pw {
         let bw: libc::c_int = imin(
             32 as libc::c_int,
             (pw as libc::c_int as libc::c_uint).wrapping_sub(bx) as libc::c_int,
@@ -979,10 +979,10 @@ unsafe extern "C" fn fguv_32x32xn_c(
     }
     let mut offsets: [[libc::c_int; 2]; 2] = [[0; 2]; 2];
     let mut bx: libc::c_uint = 0 as libc::c_int as libc::c_uint;
-    while (bx as libc::c_ulong) < pw {
+    while (bx as size_t) < pw {
         let bw: libc::c_int = imin(
             32 as libc::c_int >> sx,
-            pw.wrapping_sub(bx as libc::c_ulong) as libc::c_int,
+            pw.wrapping_sub(bx as size_t) as libc::c_int,
         );
         if (*data).overlap_flag != 0 && bx != 0 {
             let mut i_0: libc::c_int = 0 as libc::c_int;
